backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
arxiv==2.1.0
anthropic==0.34.2
python-multipart==0.0.9
google-genai

